Fix: 데이터소스탭 / list item 컴포넌트 분리
